Fix database configuration
Migrated the PostgreSQL database to use the Postgres Operator, which
manages the database deployment and provides a secret with the database
connection information. Updated the Augur deployment to use the secret for
database connection. Removed the old PostgreSQL deployment and related
configuration files. Added an example secret file for Augur configuration.

Signed-off-by: John Strunk <jstrunk@redhat.com>