Add valkey
Added valkey configuration and deployment files to enable valkey in the
augur environment. This includes a StatefulSet for valkey, a ConfigMap for
valkey configuration, and services for valkey and valkey-sentinel. The
valkey StatefulSet is configured to run with 3 replicas and uses a headless
service for DNS discovery among pods. The valkey-config ConfigMap defines
the valkey configuration, including persistence and sentinel settings.

Signed-off-by: John Strunk <jstrunk@redhat.com>